+++ src/main/kotlin/ru/yandex/money/tools/grafana/dsl/metrics/MetricsBuilder.kt
@@ -1,5 +1,6 @@
 package ru.yandex.money.tools.grafana.dsl.metrics
 
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings
 import ru.yandex.money.tools.grafana.dsl.DashboardElement
 import ru.yandex.money.tools.grafana.dsl.datasource.Datasource
 import ru.yandex.money.tools.grafana.dsl.datasource.Zabbix
@@ -11,11 +12,20 @@ import ru.yandex.money.tools.grafana.dsl.panels.graph.display.seriesoverrides.Se
 class MetricsBuilder<DatasourceT : Datasource> {
 
     val metrics = mutableListOf<DashboardMetric>()
-
     internal val seriesOverrides: MutableList<SeriesOverride> = mutableListOf()
+    private val metricIdGenerator by lazy { MetricIdGenerator() }
 
-    fun metric(referenceId: String, hidden: Boolean = false, fn: () -> Metric) {
-        metrics += ReferencedDashboardMetric(fn(), referenceId, hidden)
+    fun metric(referenceId: String? = null, hidden: Boolean = false, fn: () -> Metric) {
+        metrics += ReferencedDashboardMetric(fn(), referenceId ?: generateMetricId(), hidden)
+    }
+
+    @SuppressFBWarnings("BC_BAD_CAST_TO_ABSTRACT_COLLECTION", "RCN_REDUNDANT_NULLCHECK_OF_NONNULL_VALUE")
+    private fun generateMetricId(): String {
+        var generatedId: String
+        do {
+            generatedId = metricIdGenerator.nextMetricId()
+        } while (metrics.map { (it as ReferencedDashboardMetric).id }.contains(generatedId))
+        return generatedId
     }
 
     fun MetricsBuilder<Zabbix>.metricsQuery(build: ZabbixMetric.Builder.Metric.() -> Unit = {}) {